requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-pptx>=0.6.23
pydantic>=2.5.3
anthropic>=0.40.0
python-dotenv>=1.0.0
